--- src/nplinker/metabolomics/file_mapping_loader_base.py
@@ -1,8 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class FileMappingLoaderBase(ABC):
-
-    @abstractmethod
-    def mapping(self) -> dict[int, list[str]]:
-        ...
