--- src/nplinker/metabolomics/molecular_family_loader_base.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-from collections.abc import Sequence
-
-from nplinker.metabolomics.molecular_family import MolecularFamily
-
-
-class MolecularFamilyLoaderBase(ABC):
-    
-    @abstractmethod
-    def families(self) -> Sequence[MolecularFamily]:
-        ...
