--- src/nplinker/metabolomics/spectrum_loader_base.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-from collections.abc import Sequence
-
-from nplinker.metabolomics.spectrum import Spectrum
-
-
-class SpectrumLoaderBase(ABC):
-    
-    @abstractmethod
-    def spectra(self) -> Sequence[Spectrum]:
-        ...
