+++ tests/test_gnps_file_mapping_loader.py
@@ -1,6 +1,6 @@
 import pytest
 
-from nplinker.metabolomics.gnps_file_mapping_loader import GNPSFileMappingLoader
+from nplinker.metabolomics.gnps.gnps_file_mapping_loader import GNPSFileMappingLoader
 from tests import DATA_DIR
 
 @pytest.fixture
